benji_bot/requirements.txt
streamlit==1.31.0
streamlit-authenticator==0.3.1
yfinance==0.2.36
vaderSentiment==3.3.2
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
python-telegram-bot==20.7
praw==7.7.1
python-dotenv==1.0.0
bcrypt==4.1.2
